feat: veteriner id'ye gore aktivate
